# Optimize `_is_collection?` method